+++ objects/power/fu_alternatorgenerator/fu_alternatorgenerator.animation
