+++ objects/power/fu_alternatorgenerator/fu_alternatorgenerator.config
@@ -3,7 +3,7 @@
     "background" : {
       "type" : "background",
       "fileHeader" : "/interface/chests/chestheader.png",
-      "fileBody" : "/objects/crafting/fu_alternatorgenerator/fu_alternatorgenerator_ui.png",
+      "fileBody" : "/objects/power/fu_alternatorgenerator/fu_alternatorgenerator_ui.png",
       "fileFooter" : "/interface/chests/chestfooter.png"
     },
     "close" : {
