+++ objects/power/fu_alternatorgenerator/fu_alternatorgenerator.object
@@ -5,7 +5,7 @@
   "category" : "wire",
   "objectType" : "container",
   "price" : 200,
-  "description" : "A wood, coal or energiwood burning power source. Max of 4W power.",
+  "description" : "A wood, coal or energiwood burning power source. Max of 4W power. ^blue;Left Input^reset;: Power Switch",
   "shortdescription" : "^cyan;Alternator Generator^white;",
   "race" : "generic",
   "printable" : false,
@@ -28,10 +28,10 @@
     "base" : "fu_alternatorgenerator_base.png"
   },
   "animationPosition" : [0, 0],
-  "soundEffect" : "/objects/crafting/fu_alternatorgenerator/fu_alternatorgenerator.wav",
+  "soundEffect" : "/objects/power/fu_alternatorgenerator/fu_alternatorgenerator.wav",
   
   "slotCount" : 1,
-  "uiConfig" : "/objects/crafting/fu_alternatorgenerator/fu_alternatorgenerator.config",
+  "uiConfig" : "/objects/power/fu_alternatorgenerator/fu_alternatorgenerator.config",
   "frameCooldown" : 67,
   "autoCloseCooldown" : 3600,
