+++ objects/power/fu_alternatorgenerator/fu_alternatorgenerator_fans.frames
