+++ objects/power/fu_alternatorgenerator/fu_alternatorgeneratoricon.png
