+++ objects/power/fu_quantumgenerator/fu_quantumgenerator.object
@@ -5,7 +5,7 @@
   "category" : "wire",
   "objectType" : "container",
   "price" : 2500,
-  "description" : "Unmatched clean energy up to 128W. Fuel type will decide power output.",
+  "description" : "Unmatched clean energy up to 128W. Fuel type will decide power output. ^blue;Left Input^reset;: Power Switch",
   "shortdescription" : "^cyan;Quantum Reactor^white;",
   "race" : "generic",
   "printable" : false,
