+++ objects/power/isn_fissionreactornew/isn_fissionreactornew.object
@@ -5,7 +5,7 @@
   "category" : "wire",
   "objectType" : "container",
   "price" : 11000,
-  "description" : "Uses radioactive materials to generate power. Can generate from 4W to 80W. Higher power will be ^red;highly radioactive^reset;.",
+  "description" : "Uses radioactive materials to generate power. Can generate from 4W to 80W. Higher power will be ^red;highly radioactive^reset;. ^blue;Left Input^reset;: Power Switch",
   "shortdescription" : "^cyan;Fission Reactor^white;",
   "race" : "generic",
   "printable" : false,
@@ -43,8 +43,8 @@
   "scripts" : [ "/scripts/power.lua","isn_fissionreactor.lua" ],
   "scriptDelta" : 60,
 
-  "inputNodes" : [ [2, 4] ],
-  "outputNodes" : [ [2, 2] ],
+  "inputNodes" : [ [5, 2], [6,2]],
+  "outputNodes" : [ [4, 2] ],
   "isn_powerSupplier" : true,
   "powertype" : "output",
