+++ objects/power/isn_thermalgenerator/isn_thermalgenerator.object
@@ -5,7 +5,7 @@
   "category" : "wire",
   "objectType" : "container",
   "price" : 3500,
-  "description" : "A fuel-burning generator. It can generate roughly 10W or higher based on fuel type used.",
+  "description" : "A fuel-burning generator. It can generate roughly 10W or higher based on fuel type used. ^blue;Left Input^reset;: Power Switch",
   "shortdescription" : "^cyan;Combustion Generator^white;",
   "race" : "generic",
   "printable" : false,
