+++ objects/power/isn_thermalgenerator/isn_thermalgenerator_UNUSED.lua
